chore: reorganize src/util, mpv-* scripts, and macOS packaging assets
Group platform utils under src/util/, prefix mpv scripts by role (mpv-osc-*, mpv-media-keys, mpv-screensaver, mpv-slideshow-redraw), move rpi run-local script under scripts/rpi-*, and move the macOS app icon from assets/images/ into packaging/macos/ alongside the rest of the macOS packaging inputs. No behavior changes; CMakeLists and references updated to match new paths.
