+++ packaging/macos/240-mp.icon/icon.json
