+++ scripts/mpv-osc-ambient.lua
