+++ scripts/mpv-slideshow-redraw.lua
