+++ src/input/InputManager.h
@@ -118,7 +118,7 @@ private:
     QString labelForButton(int button) const;
     static int qtKeyForAction(Action a);
     static QString mpvKeyForAction(Action a);
-    // Maps a HID media-key event to the canonical mpv key name media-keys.lua
+    // Maps a HID media-key event to the canonical mpv key name mpv-media-keys.lua
     // binds, or an empty string for non-media keys.
     static QString mpvKeyForMediaEvent(const QKeyEvent *ke);
     static Action actionFromString(const QString &name, bool *ok);
